% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_deps.R
\name{check_deps}
\alias{check_deps}
\title{Check dependencies}
\usage{
check_deps(lesson)
}
\arguments{
\item{lesson}{name of lesson}
}
\value{
dependencies
}
\description{
Check dependencies
}
\examples{
check_deps("import")
}
