% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_lesson.R
\name{use_lesson}
\alias{use_lesson}
\title{Install lesson materials}
\usage{
use_lesson(lesson, destdir = NULL)
}
\arguments{
\item{lesson}{The name of the lesson}

\item{destdir}{The path on your computer where you would like the learning
materials installed. If \code{NULL}, defaults to your Desktop or some other
conspicuous place.}
}
\description{
\code{use_lesson()} will install the materials for a given learning lesson. Then,
it will open a new RStudio Project containing the files you'll need for that
lesson.
}
